refactor(frontend-ir): modularize evidence assembly out of frontend_ir_report (#3119)
frontend_ir_report.py mixed CLI, orchestration, and every evidence-family
assembler in one 853-line module. Continue the existing sibling-module
pattern (frontend_ir_common/_sources/_resources/_tokens/_tweaks/_proofs/
_validation) by extracting the remaining evidence helpers into dedicated
frontend_ir_* modules so the report module becomes a thin orchestrator
(build_frontend_ir + main + repo_relative).

New modules:
- frontend_ir_routes.py: route-manifest adapters + the shared low-level
  primitives (route_name, source_kind, normalize_source_of_truth,
  semantic_role, support_for_route, route_rows, routes_present,
  row_node_id, inline_source_audit, route_counts, primitive_counts,
  routes_from_rows). Owns the primitives so styles/state/nodes/sources
  import one-way from here, avoiding an import cycle.
- frontend_ir_styles.py: style_for_row, iter_style_values, style_counts.
- frontend_ir_state.py: state_for_row, state_counts.
- frontend_ir_nodes.py: nodes_from_rows.

Extended modules:
- frontend_ir_sources.py: count_map, dynamic_risks, source_span, source_spans.
- frontend_ir_tokens.py: token_key, tokens_from_rows, token_counts.
- frontend_ir_tweaks.py: tweak_counts.
- frontend_ir_resources.py: artifact_ref_from_manifest, binary_dependency_evidence.

Function bodies are moved verbatim; readiness/policy (frontend_ir_gate.py)
is untouched. frontend_ir_report.py re-imports every moved name so the
existing importlib-based test that accesses functions as fir.<name>
stays green, and so the module's public namespace is unchanged.

Behavior-preserving: report JSON is byte-identical before and after
(verified via a build_frontend_ir + CLI snapshot harness; sha256
17be8225...c5462927 matches across the move). All 84 pre-existing
frontend_ir tests stay green.
